Split withdrawal
