+++ src/newClient/types/transfer.ts
@@ -34,12 +34,3 @@ export interface PaymentResponse {
   amount: string;
   status: string;
 }
-
-export interface WithdrawRequest {
-  address_id: string;
-  amount: string;
-  trace_id?: string;
-
-  memo?: string;
-  pin?: string;
-}
