+++ src/newClient/types/withdrawal.ts
@@ -1,0 +1,8 @@
+export interface WithdrawRequest {
+  address_id: string;
+  amount: string;
+  trace_id?: string;
+
+  memo?: string;
+  pin?: string;
+}
